+++ app/components/assets/form.tsx
@@ -36,6 +36,7 @@ import {
   TooltipTrigger,
 } from "../shared/tooltip";
 import { TagsAutocomplete } from "../tag/tags-autocomplete";
+import { scrollToError } from "~/utils/scroll-to-error";
 
 export const NewAssetFormSchema = z.object({
   title: z
@@ -118,41 +119,7 @@ export const AssetForm = ({
   }>();
 
   const handleSubmit = () => {
-    const scrollToFirstError = () => {
-      const errorElements = document.querySelectorAll(".text-error-500");
-      if (errorElements.length > 0) {
-        let firstError = errorElements[0];
-        const elementToScrollTo =
-          firstError.previousElementSibling || firstError;
-        elementToScrollTo.scrollIntoView({ behavior: "smooth" });
-      }
-    };
-
-    scrollToFirstError();
-
-    const observer = new IntersectionObserver(
-      (entries, observer) => {
-        const notIntersecting = entries.filter(
-          (entry) => !entry.isIntersecting
-        );
-
-        if (notIntersecting.length > 0) {
-          notIntersecting.forEach((entry) => {
-            const elementToScrollTo =
-              entry.target.previousElementSibling || entry.target;
-            elementToScrollTo.scrollIntoView({ behavior: "smooth" });
-            observer.unobserve(entry.target);
-          });
-        }
-      },
-      { threshold: 0.1 }
-    );
-
-    const target = document.body;
-
-    return () => {
-      observer.disconnect();
-    };
+    scrollToError();
   };
 
   return (
